refactor: PR review fixes for negative prompts
- Fix SAM3 AttributeError: initialize _negative_geometry_features and
  _negative_geometry_mask in __init__
- Fix filter_by_category() dropping is_reference and n_shot fields
- Extract NegativeEmbeddingMixin to eliminate code duplication between
  Matcher and PerDino (_extract_negative_embedding, _adjust_similarities)
- Deduplicate _select/_sel helpers into Sample._index_select()
- Remove dead _negative_points from Matcher/PerDino (only SAM3 uses it)
- Replace hardcoded personal paths with env vars (PERSEG_DIR, COCO_DIR)
- Change NegativeMaskToPoints labels dtype from float32 to torch.long
- Use strict=True in SAM3 _extract_negative_points zip
- Add test for SAM3 geometry attrs safe without background masks
- Document SAM3 cross-image size limitation in class docstring
-e

+++ library/src/instantlearn/components/negative_prompts.py
@@ -82,5 +82,5 @@ class NegativeMaskToPoints(nn.Module):
         else:
             points = torch.empty(0, 2, device=negative_masks.device, dtype=torch.float32)
 
-        labels = torch.zeros(points.shape[0], device=points.device, dtype=torch.float32)
+        labels = torch.zeros(points.shape[0], device=points.device, dtype=torch.long)
         return points, labels
